feat: friend store을 따로 하지 않고 user store에 추가
