--- src/frontend/modules/friend/actions.ts
@@ -1,3 +1,0 @@
-import { createRequestActionTypes } from 'modules/utils/actionUtils';
-
-export const [GET_FRIEND, GET_FRIEND_SUCCESS, GET_FRIEND_FAIL] = createRequestActionTypes('GET_FRIEND');
