--- src/frontend/modules/friend/index.ts
@@ -1,3 +1,0 @@
-export { default } from './reducer';
-export * from './actions';
-export * from './types';
