--- src/frontend/modules/friend/initalData.ts
@@ -1,10 +1,0 @@
-export const initalFriendInfo = [
-  {
-    id: 1,
-    nickname: '[TEST]nick',
-    profile: {
-      image: '',
-      description: '[TEST]desc',
-    },
-  },
-];
