+++ src/frontend/modules/user/reducer.ts
@@ -1,12 +1,14 @@
 import { createReducer } from 'typesafe-actions';
 
 import { asyncState } from 'modules/utils/reducerUtils';
-import { initalUserInfo } from 'modules/user/initalData';
+import { initalUserInfo, initalFriendInfo } from 'modules/user/initalData';
 import { IUserState } from './types';
 import { SAVE_USERINFO, SAVE_USERINFO_SUCCESS, SAVE_USERINFO_FAIL } from './actions';
+import { GET_FRIEND, GET_FRIEND_SUCCESS, GET_FRIEND_FAIL } from './actions';
 
 const initialState: IUserState = {
   userInfo: asyncState.initial(initalUserInfo),
+  friends: asyncState.initial(initalFriendInfo),
 };
 
 const reducer = createReducer<IUserState>(initialState, {
@@ -21,6 +23,19 @@ const reducer = createReducer<IUserState>(initialState, {
   [SAVE_USERINFO_FAIL]: (state, action) => ({
     ...state,
     userInfo: asyncState.error(initalUserInfo, action.payload),
+  }),
+
+  [GET_FRIEND]: (state, action) => ({
+    ...state,
+    friends: asyncState.load(initalFriendInfo),
+  }),
+  [GET_FRIEND_SUCCESS]: (state, action) => ({
+    ...state,
+    friends: asyncState.success(action.payload),
+  }),
+  [GET_FRIEND_FAIL]: (state, action) => ({
+    ...state,
+    friends: asyncState.error(initalFriendInfo, action.payload),
   }),
 });
