+++ src/frontend/modules/user/types.ts
@@ -10,8 +10,20 @@ interface IUserInfo {
   idx: number;
   nickname: string;
 }
+
+export interface IFriendInfo {
+  id: number;
+  nickname: string;
+  profile: {
+    image: string | JSX.Element;
+    description: string;
+  };
+  is_friend?: number;
+}
+
 export interface IUserState {
   userInfo: AsyncState<IUserInfo, Error>;
+  friends: AsyncState<IFriendInfo[], Error>;
 }
 
 export type userAction = ActionType<typeof actions>;
