+++ src/frontend/pages/signin.tsx
@@ -8,6 +8,8 @@ import { IState } from 'modules';
 import { saveUserInfo } from 'modules/user';
 import { doSignInAPI } from 'api/auth/api';
 import { parseToken } from 'util/parseToken';
+import { GET_FRIEND } from 'modules/friend';
+import { getFriendsAPI } from 'api/friend/api';
 
 import Text from 'components/atoms/Text';
 import FilledButton from 'components/atoms/FilledButton';
@@ -62,6 +64,8 @@ const signin: NextPage<IState> = () => {
         const result = parseToken(res.data.accessToken);
         dispatch(saveUserInfo.request(result));
         router.push('/category'); // TO DO(양하): 메인으로 redirect 변경, 지금 메인페이지 오류나서 일단 카테고리 페이지로 redirect
+
+        //친구 목록 불러와서 API에 저장
       }
     }
   };
